[refactor] Add SubscriptionSWQueryRepo to SubscriptionSWRepo
